+++ nonebot_plugin_batarot/__init__.py
@@ -1,6 +1,13 @@
-from nonebot import on_command
+from nonebot import require
 from nonebot.plugin import PluginMetadata
-from .commands import handle_tarot, handle_tarot_spread, handle_daily_fortune, handle_tarot_reading
+
+"""
+加载saa插件 提供多适配器支持
+:::notice 请勿重复加载saa
+"""
+require("nonebot_plugin_saa")
+
+from nonebot_plugin_saa import __plugin_meta__ as saa_plugin_meta
 
 __version__ = "0.2.0.post3"
 __plugin_meta__ = PluginMetadata(
@@ -10,16 +17,7 @@ __plugin_meta__ = PluginMetadata(
     homepage="https://github.com/Perseus037/nonebot_plugin_batarot",
     type="application",
     config=None,
-    supported_adapters={"~onebot.v11"},
+    supported_adapters=saa_plugin_meta.supported_adapters,
 )
 
-
-tarot = on_command("batarot", aliases={"ba塔罗牌"})
-tarot_spread = on_command("divination", aliases={"ba占卜"})
-tarot_fortune = on_command("fortune", aliases={"ba运势"})
-tarot_reading = on_command("reading", aliases={"ba塔罗牌解读"})
-
-tarot.handle()(handle_tarot)
-tarot_spread.handle()(handle_tarot_spread)
-tarot_fortune.handle()(handle_daily_fortune)
-tarot_reading.handle()(handle_tarot_reading)
+from . import handler as handler
